fix: LCP breakdown
